+++ NEWS.md
@@ -2,6 +2,8 @@
 
 **New features**
 
+- new function `to_gt()` for converting the result of `look_for()` into a nicely
+  formatted table (#189)
 - new `sep` argument for `names_prefixed_by_values()`
 - new arguments (`sep_value_labels` and `sep_other`) for
   `convert_list_columns_to_character()` and `lookfor_to_long_format()`