+++ NEWS.md
@@ -2,11 +2,12 @@
 
 **New features**
 
+- new function `to_gt()` for converting the result of `look_for()` into a nicely
+  formatted table (#189)
 - new functions `dictionary_to_variable_labels()` and
   `dictionary_to_value_labels()` to convert a dictionary data frame into a list
   of variable / value labels (#183)
-- new function `to_gt()` for converting the result of `look_for()` into a nicely
-  formatted table (#189)
+- new `.overwrite` argument for `set_variable_labels()` (#183)
 - new `sep` argument for `names_prefixed_by_values()`
 - new arguments (`sep_value_labels` and `sep_other`) for
   `convert_list_columns_to_character()` and `lookfor_to_long_format()`